requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aioredis==1.3.1
aniso8601==7.0.0
arq==0.19
async-timeout==3.0.1
attrs==20.1.0
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.26.1
graphene==2.1.8
graphene-mongo==0.2.13
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.16
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0
httptools==0.1.1
idna==2.10
ipaddress==1.0.23
iso8601==0.1.12
joblib==0.15.1
mongoengine==0.20.0
mongomock==3.20.0
monotonic==1.5
multidict==4.7.6
mypy==0.782
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.0
pandas==1.0.5
promise==2.3
pycares==3.1.1
pycparser==2.20
pydantic==1.6.1
pyflakes
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
rq
Rx==1.6.1
sentinels==1.0.0
singledispatch==3.4.0.3
six==1.15.0
sphinx
tqdm==4.46.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.7
uvloop==0.14.0
websocket-client==0.57.0
websockets==8.1
yarl==1.5.1
